test: Move "common" and "sample" code used by tpmclient to tpmclient.
Since the resourcemgr is gone this code is no longer "common". This
change moves this code closer to the only place it's still used:
tpmclient.


Conflicts:
	Makefile.am
	bootstrap
	common/sockets.c
	common/sockets.cpp
	sysapi/sysapi_util/checkoverflow.c
	tcti/sockets.c
	tcti/tcti_socket.cpp
	test/tpmclient/tpmclient.cpp
