+++ bootstrap
@@ -26,10 +26,6 @@ echo "Generating file lists: ${VARS_FILE}"
   src_listvar "common" "*.c" "COMMON_C"
   src_listvar "common" "*.h" "COMMON_H"
   printf "COMMON_SRC = \$(COMMON_C) \$(COMMON_CXX) \$(COMMON_H)\n"
-
-  src_listvar "test/common/sample" "*.c" "SAMPLE_C"
-  src_listvar "test/common/sample" "*.h" "SAMPLE_H"
-  printf "SAMPLE_SRC = \$(SAMPLE_C) \$(SAMPLE_H)\n"
 ) > ${VARS_FILE}
 
 ${AUTORECONF} --install --sym
