+++ common/debug.h
@@ -31,7 +31,6 @@
 #include "sapi/tpm20.h"
 #include "tcti/tcti_socket.h"
 #include <stdio.h>
-#include "sockets.h"
 
 #ifdef __cplusplus
 extern "C" {
