+++ tcti/platformcommand.c
@@ -46,6 +46,7 @@
 #include "sysapi_util.h"
 #include "common/debug.h"
 #include <sapi/tss2_tcti.h>
+#include "sockets.h"
 #include "tcti_util.h"
 #include "logging.h"
