+++ test/tpmclient/DecryptEncrypt.c
