+++ test/tpmclient/StartAuthSession.c
