+++ test/tpmclient/TpmHandleToName.c
