+++ test/tpmclient/TpmHmac.c
