+++ test/tpmclient/tcti_util.h
